Add consistent scaling
Previously square-root scaling was expected inside the object
identification algorithm, but wasn't applied on the input (instead a
linear scaling was used). Scaling has been added to a separate routine
and bounds are checked on input to avoid overflow on the short which is
used for storage of scaled values.

TODO: Scaling hasn't been applied to main cli yet.